Removing old code
Confirmed that "deprecated" code was intended to be removed for
good, so I removed it from the tree, as well as the apps that
depended on it. Also removed the "hold" directory (and contained
code) from the CRF library module. I suspect that it stuck around
because of SVN abuse.

Removed code will continue to be available in the original SVN
repo, if needed (it will also remain in the history of this repo,
of course).