Update README; conditional demo install; update demo script
Updated documentation for end-user installation. Changed
installation of demonstration scripts to depend on the existence
of the Kaldi parameter for the configure script.

Modified demo script to use pushd/popd instead of cd, which would
have given screwy results. Also added a check for command line
args and usage statement.